requirements.txt
aiohttp==3.8.5
lxml==4.9.3
numpy==1.24.3
opencv_python==4.8.1.78
pycryptodomex==3.19.0
requests==2.31.0
tabulate==0.8.10
